requirements.txt
aiogram==3.0.0rc1
Django==2.2.19
python-dotenv==1.0.0
pytz==2022.1
sqlparse==0.4.2
Pillow==8.3.1
flake8==6.1.0
isort==5.12.0
aiogoogle==5.4.0
djangorestframework==3.10.3